requirements.txt
torch==2.5.1
torchvision==0.20.1
numpy==1.26.3
roma
gradio==5.17.1
pydantic==2.10.6
matplotlib
tqdm
opencv-python
pypng
scipy
einops
trimesh
pyglet<2
viser
jaxtyping
hydra-submitit-launcher
scikit-learn
plotly
git+https://github.com/facebookresearch/vggt.git@44b3afb
